Extract intendant-platform: platform.rs, vision.rs, DisplayTarget
Workspace-extraction lap 2: the two OS-integration leaves leave the
caller bin. crates/intendant-platform holds platform.rs (process
liveness/cmdline probes, spawn helpers; the windows-sys dependency
moves with it — platform.rs was its only user in the root package),
vision.rs (virtual-display lifecycle), and DisplayTarget (hoisted from
computer_use.rs, which now re-exports it at the old mount point). The
caller mounts both modules with use-aliases so every existing
crate::platform / crate::vision path keeps compiling; four vision
helpers widen pub(crate) -> pub for their cross-crate callers, and
vision's tests get a crate-local env lock.

Two latent problems surfaced and fixed along the way:

- CI's cargo test/check lines applied --bins globally, which silently
  no-ops packages that have no bin targets: intendant-core and
  intendant-display lib tests (~570) have not been running in CI since
  their extraction. windows.yml now uses one invocation for the root
  bins and a separate one for the lib crates (which this change also
  proves green: core 71, display 499, platform 27 on Linux).
- vision's find_free_display test probed the real /tmp X11 locks and
  could reach the kill-and-reclaim path — on a box with a live :99
  (the Linux CI runner is one) it fails, and same-user it would kill a
  real display. find_free_display gains a lock-dir-injectable core and
  the test pins a temp dir.
