+++ crates/intendant-platform/Cargo.toml
@@ -1,0 +1,33 @@
+[package]
+name = "intendant-platform"
+version = "0.1.0"
+edition = "2021"
+license = "MIT"
+publish = false
+
+[dependencies]
+intendant-core = { path = "../intendant-core" }
+serde = { version = "1.0", features = ["derive"] }
+tokio = { version = "1.0", features = ["process", "sync", "io-util", "macros", "rt"] }
+# PATHEXT-aware executable resolution for `platform::spawn_command` (npm
+# `.cmd`/`.bat` shims on Windows, which CreateProcess does not resolve).
+which = "8"
+dirs = "6"
+
+[dev-dependencies]
+# platform's spawn/lock tests write into temp dirs.
+tempfile = "3"
+
+[target.'cfg(unix)'.dependencies]
+libc = "0.2"
+
+[target.'cfg(windows)'.dependencies]
+windows-sys = { version = "0.59", features = [
+    "Win32_Foundation",
+    "Win32_System_Diagnostics_ToolHelp",
+    "Win32_System_Threading",
+    "Win32_System_ProcessStatus",
+    "Wdk_System_Threading",
+] }
+
+# DisplayTarget serializes on the command wire (tagged enum).
